fix: ioc tests
