--- SpaceBattle.Lib/Classes/MovingObjectAdapter.cs
@@ -1,18 +1,0 @@
-// namespace SpaceBattle.Lib;
-
-// public class MovinngObjectAdapter(IDictionary<string, object> obj) : IMovingObject
-// {
-//     private readonly IDictionary<string, object> _obj = obj;
-
-//     public Vector Position
-//     {
-//         get => (Vector)obj["Position"];
-//         set => obj["Position"] = value;
-//     }
-
-//     public Vector Velocity
-//     {
-//         get => (Vector)obj["Velocity"];
-//         set => obj["Velocity"] = value;
-//     }
-// }
