--- SpaceBattle.Tests/EmptyCommandTest.cs
@@ -1,13 +1,0 @@
-using SpaceBattle.Lib;
-
-namespace SpaceBattle.Tests;
-
-public class EmptyCommandTest
-{
-    [Fact]
-    public void EmptyCommandInitPositive()
-    {
-        var emptyCommand = new EmptyCommand();
-        emptyCommand.Execute();
-    }
-}
